requirements.txt
streamlit 
autogen 
python-dotenv 
groq 
fpdf